Fix id used with Long type in app and Integer in db

+++ BackEnd/b2formeditor/src/main/java/com/b2formeditor/repositories/FormRepository.java
@@ -11,12 +11,12 @@ import java.util.List;
 public interface FormRepository extends MongoRepository<Form, Integer> {
     List<Form> findAll();
 
-    Form findOne(Long id);
+    Form findOne(Integer id);
 
     List<Form> findByCreated_by(String created_by);
 
     List<Form> findByTitle(String title);
 
-    void delete(Long id);
+    void delete(Integer id);
 
 }
